Revert "Issue #2870458 by Lendude, jibran, alexpott, vijaycs85, Wim Leers, dawehner, larowlan: Convert web tests to browser tests for quickedit module"
This reverts commit 2871d43e0d4c025f2533f77280d56f5f2fda30ad.
